37928 - added multi tenant to events and covered actions with tests

+++ paypal-commercetools-events/src/connector/actions.ts
@@ -4,6 +4,7 @@ import {
   deleteAccessToken,
   getCachedAccessToken,
 } from '../service/config.service';
+import { logger } from '../utils/logger.utils';
 
 const PAYPAL_PARCEL_ADDED_TO_DELIVERY_KEY = 'paypal-parcelAddedToDelivery';
 
@@ -59,8 +60,17 @@ export async function deleteParcelAddedToDeliverySubscription(
   }
 }
 
-export const deleteAccessTokenIfExists = async () => {
-  if (await getCachedAccessToken()) {
-    await deleteAccessToken();
+export const deleteAccessTokenIfExists = async (
+  lastTenantType?: 'single' | 'multi'
+) => {
+  if (lastTenantType === 'single' || !lastTenantType) {
+    if (await getCachedAccessToken()) {
+      await deleteAccessToken();
+    }
+  }
+  if (lastTenantType === 'multi' || !lastTenantType) {
+    if (await getCachedAccessToken(true)) {
+      await deleteAccessToken(true);
+    }
   }
 };
